✨ feat: 퀴즈 아이템 틀 잡고 헤더 일단 만듬
