+++ services/quiz/package.json
@@ -22,6 +22,7 @@
     "@emotion/react": "^11.9.3",
     "@emotion/styled": "^11.9.3",
     "@toss/assert": "^1.1.0",
+    "@toss/emotion-utils": "^1.1.3",
     "@toss/error-boundary": "^1.1.0",
     "@toss/react": "^1.3.0",
     "@toss/react-query": "^1.1.0",
