+++ services/quiz/pages/study-detail/[studyId].tsx
