+++ services/quiz/src/components/StudyDetailPage/QuizItem/style.ts
@@ -1,11 +1,25 @@
-// import { theme } from '@depromeet/theme';
+import { theme } from '@depromeet/theme';
 import styled from '@emotion/styled';
 
 export const Item = styled.li`
+  width: 982px;
+  padding: 22px 38.5px 22px 28px;
+  border-radius: 12px;
+  :nth-child(n+2) {
+    margin-top: 12px;
+  }
 `
 
-export const ItemHeader = styled.summary`
+export const ItemDetails = styled.details`
   
+`;
+
+export const ItemHeader = styled.summary`
+  display: flex;
+  list-style: none;
+  align-items: center;
+
+
 `;
 
 export const ItemBody = styled.div`
