+++ services/quiz/src/components/StudyDetailPage/QuizList/style.ts
@@ -1,0 +1,7 @@
+// import { theme } from '@depromeet/theme';
+import styled from '@emotion/styled';
+
+export const List = styled.ul`
+    list-style: none;
+    padding: 0;
+`;
